+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for pod security policy in kubernetes
 - Support for Oracle Big Data Service
+- Support for application definition parameters update in dataflow application
 
 ## 3.68.0 (March 25, 2020)